requirements.txt
asgiref==3.9.1
Django==4.2.7
django-ckeditor==6.7.0
django-js-asset==3.1.2
Pillow>=11.0.0
sqlparse==0.5.3
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django-debug-toolbar==4.3.0
